virtio-scsi: Add wwpn and tgpt properties
A vhost-scsi device must be bound to a tcm_vhost target port.  This will
be done by issuing an ioctl to the device with the wwpn and tgpt.  This
patch adds these properties to the virtio-scsi-pci device:

  qemu -device virtio-scsi-pci,wwpn=$TARGET_WWN,tgpt=$TPGT

Future patches will open /dev/vhost-scsi and issue the ioctl.

This patch also moves the virtio_scsi_init() prototype to
hw/virtio-scsi.h where the VirtIOSCSIConf structure is defined.  I think
this is cleaner than defining it in hw/virtio.h.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>
Signed-off-by: Zhi Yong Wu <wuzhy@linux.vnet.ibm.com>